Add dpdk as a git submodule pointing at tag v1.8.0
This is so we don't have to track the entire dpdk source tree in the onvm repository, but can instead reference another repository at a certain commit. This keeps the repository cleaner and makes the dependencies more decoupled. It'll also make updates easier, since we just have to update the pointer in the submodule.

This commit contains three steps:
 (1) Remove dpdk from being tracked in this repo
 (2) Add dpdk submodule at v1.8.0 (https://github.com/sdnfv/onvm-dpdk/releases/tag/v1.8.0)
 (3) Update docs to init submodule - (e.g. just run `git submodule init && git submodule update` to pull changes in)