Shutdown NFs by passing message to manager
This commit fleshes out the lifecycle of an NF by adding a few more states in `common.h`.
 1. WAITING_FOR_ID: from the time an NF starts up until the manager processes its info struct
 2. STARTING: the rest of the init process, before the NF is ready to receive packets
 3. RUNNING: same as before
 4. PAUSED: unused currently, but may be useful (could be an extension to buffering - a "stop sending me packets for a bit" flag)
 5. STOPPED: when the NF is engaged in the exit process

I had to split up WAITING_FOR_ID and STARTING due to a race condition that occurs now that we check the status in the info struct queue processor in the manager
(this is to differentiate starting and stopping NFs). In the case where a NF sets its own ID, it won't wait for the manager to process its item so it'll very
quickly set its own status to RUNNING, thus not doing anything by the time the manager polls its queue and checks the status to see if it's starting or stopping.
Now an NF will always wait for the manager to process its info struct to avoid this kind of error.

This commit, in the process of cleaning up NF start/stop code, also fixes #85 in the process. It simply always stores an NF's info struct at the same index as its
ID and also ensures to set the `client_id` field in the client struct, which is used in some other places in the code.