Updating Pktgen-DPDK Documentation (#176)
This updates the documentation of Pktgen-DPDK and adds a submodule to our fork of it [onvm-pktgen](https://github.com/sdnfv/onvm-pktgen).  We have removed the configuration file and run script from this repo and have moved them into the Pktgen repo.  This submodule of Pktgen points to and uses a version compliant with DPDK v16.11

Commit log:

* Updating pktgen docs, sh/lua config files
* Increased pkt count in pktgen-config
* Updated docs to fix pcap error on install
* Documentation formating
* Updated pktgen docs
* Update pktgen documentation
* Update pktgen doc
* Added onvm-pktgen submodule
* Modified name, repo for pktgen submodule
* Revert to 5efccda
* Added pktgen submodule at tools/Pktgen
* Moved and renamed pktgen.md to tools/Pktgen
* Update pktgen doc
* Update pktgen doc
* Update run-pktgen path
* Removed pktgen-dpdk directory from scripts, moved scripts to onvm/pktgen
* Update pktgen doc
* Update docs to match new script folder
* Updating onvm-pktgen submodule pointer
* Removed unnecessary instructions in docs