Bug Fix: Fixes Speed Tester Port Hack (#189)
Addressing a contentious bug in the Speed Tester NF where fake packets were marked as coming in through NIC port 3.  This is bad since a host can have more than three NIC ports.

This addresses the bug by creating a new field in the `onvm_meta` struct for NF packet flags.  There are now 8 bits that a NF developer can toggle to store additional information in a packet.  In the file onvm/onvm_common.h, we provide a set of macro functions to facilitate using the flags field.

It is the NF's responsibility to make sure these flags are not used in a contentious way.  This means that the ONVM manager does not operate on the flags field.  

Commit log:

* Bug Fix: Fixes Speed Tester Port Hack (#181)

Improves the way that 'speed_tester' marks the packets it generates. Before, fake packets would come through port 3. With the updated code, a flag in the 'onvm_pkt_meta' struct indicates when a packet is fake.

* Bug Fix: Fixes Speed Tester Port Hack (#181)

Adds byte flag to 'onvm_pkt_meta' called 'fake_packet' which indicates a packet generated by 'speed_tester'.

* Implemented requested changes

Added macro function, changed `fake_packet` to `flags`, and removed line that sets packet's port.

* Implemented requested changes - flag name changed

Fixed the spacing and changed name of field to `flags`. Additionally, added comment suggested by @twood02

* Bug Fix: Fixes Speed Tester Port Hack (#181)

Added macro functions for checking and maniplulating individual bits in `onvm_pkt_meta` flags.

* Implemented `ONVM_SET_BIT` macro function in `speed_tester`.

* Fixed spacing.