[Bug fix] Implement Three Step NF Startup Handshake (#202)
Previously, if an NF was killed before it received an ID from the manager, the manager would not be aware of this.  When the manager attempted to exit, it would infinitely wait and attempt to terminate this NF that isn't running anymore.

This is now fixed by implementing a three step handshake for NFs when they register themselves with the manager.  The manager will not increment NF state variables until it receives `MSG_NF_READY` from an NF.  It also limits how many times the manager will attempt to kill NFs upon exit.

API Changes:

>when building an NF that uses advanced rings, be sure to call the `onvm_nflib_nf_ready` function to ensure this message gets sent.

Closes #184 

Commit log:

* Add NF message to signal startup

As described in #184

* Cap the number of iterations in shutdown loop

As described in https://github.com/sdnfv/openNetVM-dev/issues/184

* Ensure we call nf_ready in sequence