Manager Stats Thread SleepTime Argument (#204)
Before, the stats thread would sleep every 1 second and print out information about NIC ports and NFs then.  Now, you can pass an argument to ONVM and set the sleep time to whatever the user wants.  If no sleep time is specified, then ONVM will default to 1 second.

This change modifies the argument parsing logic for ONVM manager, and sleeping in the stats thread.  This feature is useful for people trying to take full advantage of the ONVM web console.  A user can now throttle the refresh rate there with more control by also adjusting the sleep time for the ONVM manager.  Every time the manager stats thread runs, stats information is added to a JSON file which the web console reads and displays information from.  Configuring the rate at which ONVM outputs this information enables tighter control on the web console side.

Commit log:

* Added ability to specify a sleep time with -z.

* Fixed default behavior for stats sleep time.

Made it so default behavior is 1 second delay.  Prior to this commit,
it had a default delay of 0 seconds, which is not desired behavior.

* Made requested changes.