NF Drop Rates Displayed In Stats (#206)
This PR adds NF packet drop rates to the stats output.  A small change was made to the ONVM manager to calculate and then output NF drop rates.  This information was then appended to the JSON objects that the web console uses to display information.

On the web console, the drop rate is displayed beside the graph title for each NF.  This information is _not_ a line on the graph because drop rates are _not_ calculated for the ONVM host NIC ports yet.  The way graphs are generated for the web console requires that graph data be the same for _both_ NF and Port information.  Since this is not the case, the NF drop rate is not a line on the graph, rather displayed next to the graph title.

Commit log:

* Added drop rate stats for NFs in manager and on frontend web console.

* Fixed issue when dealing with ports