Adds Script To Analyze ONVM Stats (#201)
Adds a script to aid in analyzing and understanding ONVM statistics.  Using the web console for stats, a user has the option to download a CSV file containing a log of statistics for each NF and NIC Port that ONVM managed during the runtime of the web console.  This CSV file contains raw data for each NF and port and is hard to consume and understand.  This script parses the data, and prints out a summary of statistics for each NF and each Port in a user-friendly manner.

Commit log:

* Created script to analyze web csv download.