[Bug Fix] Flow Table Iteration and Key Hashing Fixes (#211)
Previously, iterating a flow table did not return the new data, it only updated the index.  Now, the index is used to retrieve the data at that entry in the table.

`onvm_ft_remove_key()` did not call `onvm_softrss()` to calculate the proper RSS hash to be used for insertion/deletion from the flow table.  Previously, `onvm_ft_remove_key()` would use the DPDK hashing function which would result in incorrect deletions from the flow table.

It is good to note that any flow table function that uses keys to operate on the table (`onvm_ft_*_key()`) should always call `onvm_softrss()` to ensure usage of the proper RSS hash key.  Functions that operate on the flow table through packets (`onvm_ft_*_pkt()`) do not need to do this since the RSS hash is contained in the packet itself.

Commit log:

* [Bug Fix] Update Data Pointer in ONVM Flow Table
- onvm_ft_iterate used to correctly call rte_hash_iterate, but would not update
data pointer to reflect the returned key
- The function now calls rte_hash_iterate, and uses the returned index to retrieve
the associated data entry

* Bug Fix: Fixes Key Removal from ONVM Flow Table

-onvm_ft_remove_key utilized the built-in DPDK hashing function, causing
the key provided to be hashed to a different index than where it was
inserted
-The function now uses the onvm softrss function to obtain a hash, and
passes that to DPDK to properly remove the key from the hash table