[Bug Fix] Add Missing Arg to ONVM Go Script Getopt
Before, using the ONVM go script to run the manager with a supplied base
virtual address would fail.  This was due to a missing `v` option in the
getopt short opt list.

Simply adding `v` to the ONVM go script's getopt short opt list
addresses this.  Now running the ONVM go script with the `-v` flag works
as intended.