[Bug Fix] ONVM Install Script Handles Hugepage Config (#16)
Previously, ONVM's install script would not check `/etc/fstab` if an entry to mount the hugepage filesystem was already present and would re-add the entry each time the script was run.  Also, the install script would not "skip" configuring hugepages if the user exported the respective environment variable.  Lastly, the install guide did not document that the environment variable `ONVM_HOME` is necessary to install ONVM.

Now, the install script checks `/etc/fstab` to see if the hugepage entry is present and skips adding it if it is.  Additionally, the install script is fixed to "skip" hugepage configuration if the user exports the respective environment variable.  This is fixed by addressing a typo where a `$` was missing.  And lastly, the install guide has been amended to include notes on setting the `ONVM_HOME` environment variable to install ONVM.

Commit log:

* Add `$` so bash treats `ONVM_SKIP_HUGEPAGES` as a variable instead of a string. 

Should fix the issue where onvm manager would crash on fresh install.

* Add instruction to set `ONVM_HOME` variable.

`ONVM_HOME` is a required variable, so that should be reflected in the doc.

* Add a check to install.sh so hugefs line is added to `/etc/fstab` only once.