Flow Tracker NF (#214)
This adds a new NF, the Flow Tracker NF, which displays incoming network flow information as it forwards traffic to a specified service ID.  The NF leverages the callback handler NFLib API to user time based printing to display network flow information.  As the flow table fills up, the NF will clear expired entries and continue to operate.

In addition to the new NF, this change brings and uses a new API to the ONVM flow table library - `onvm_ft_fill_key_symmetric`.  This API ensures that bidirectional network flows are considered to be one flow instead of two.

Commit log:

* Added flow tracker NF

-This NF stores incoming flows and sends them to a specified service ID
-It will periodically print information about the contents of the table
-Allows for manually setting destination service ID and print frequency
-Stores all current flows and metadata associated with them
-Clears expired entries when table is full
-Uses 'callback_handler()' to allow for time-based printing, as opposed
to printing after a certain number of packets

* Added minor style fixes and NULL argument checks

* Modified timing system to use cpu cycles instead of getting system time

* Updated flow tracker to use onvm_fill_key_symmetric

* Added onvm_ft_fill_key_symmetric()

- This has identical functionality to onvm_ft_fill_key, but it sets the
"smaller" IP address to be the destination address
- This ensures that a bidirectional flow is considered to be one flow

* Added flow_tracker to makefile in example NF directory