Replay PCAPs with Speed Tester NF (#220)
This enables the speed_tester NF to replay PCAP alongside its normal functionality.  Previously, speed_tester would generate a buffer of "fake" packets and continuously send them to itself or a destination ID.  Now, speed_tester can be given an argument to a PCAP file and it will continuously replay it and send those packets to itself or to a destination ID.

This is useful for debugging ONVM TX core performance for more realistic workloads.  Sending the fake packets does not tell how ONVM performs under a real workload.  Replaying a PCAP file with heavy web traffic (an example) gives us more realistic readings for how an NF's TX core can handle that workload.

Directions on how to use the new speed_tester functionality are included in its README file.  This feature requires `libpcap-dev` to be installed on your host.  Since this is an _optional_ feature, this dependency is _optional_ too.  This is only required if you wish to use the PCAP replay feature.  Furthermore, there are a bunch of example PCAP files included with speed_tester too.

Along with the feature to speed_tester, this PR also increases the NF TX queue size from 128 to 16384 so NFs can store more packets in their TX queue.

Commit log:

* Adds pcap replay functionality to speed_tester NF

* Added softrss calculation

* Example test for 64B sample pcap 18/19

* Increase NF queue size

* Removed extra port configurations

* Clean up, style fixes

* Update README

* README typo fix

* Added pcap library checks

* Added requested changes, pcap samples

* Update README.md

* pcap file name fixes