[Bug Fix] Fix test_flow_dir Crash With Error Handling (#224)
This resolves #222 

Previously, the test_flow_dir NF would crash if `onvm_flow_dir_add_pkt` would return an error.  In the packet handler function for the NF, there was no error checking on the status code of that function call.  If the function failed, we would still try to do memset on a pointer that is not allocated which would cause the NF to fail and crash.  Also, the code suggested that packets with RSS hash 0 could not be handled which is not true.

Now, we check the status code of `onvm_flow_dir_add_pkt` after it is called, and drop the packet if the function failed.  This prevents the NF from crashing if the function is unable to add the packet to a flow.  This change also removes the erroneous check for packet RSS hash being 0.

Commit log:

* Error checking

* No need to check for rss=0