Added ARP Response NF (#226)
Introduces a new example NF that responds to ARP request packets for the ONVM machine's MAC addresses. This allows other machines to discover the ONVM machine based on IP across the network.

The NF takes a list of IPs as a command line argument and assigns them to the corresponding NIC ports.  The NF also takes a destination NF service ID.  Any packets that don't match an ARP request for the given IP will be forwarded to the destination NF.

This can be used in a service chain with other NFs (such as `examples/load_balancer`) which require IP-based connections.


Commit Log:
* Added ARP Response NF

Using the IP passed in at the start, the NF can reply to ARP request
packets directed at the same IP that the machine is set to.

* Modified argument input to allow for comma-separated IP addresses

* Fixes port id bug

* Added onvm_pkt_parse_ip_string() to convert a string IP to decimal

* Added struct for global NF variables and modified IP string processing

* Updated documentation and copyright information

* Revert "Added onvm_pkt_parse_ip_string() to convert a string IP to decimal"

* Modified IP processing to use function from PR #227

* Fixed IP endian check

* Implemented requested changes

* Fixed style change and arg count of strncopy()

* Removed hard-coded string length