Un-Inline onvm_pkt_enqueue_nf Function (#23)
Previously, the function `onvm_pkt_enqueue_nf` was defined as an inline function.  This function does not need to be inline - it is an API function which should not be treated as "inline".  Furthermore, newer versions of GCC complained at this function while it was inline.  Newer compilers are smart enough to detect why this does not need to be inlined.

This change removes the inline modifier from the function allowing it to compile with older and newer versions of GCC.