[Bug Fix] Flow_Table NF Sanitizes Memory During Init (#24)
If the Flow_Table NF fails to connect to an SDN controller upon startup, it returns an error state and exits.  During the exit sequence, the Flow_Table NF does not properly sanitize memory that it allocated.  Subsequent attempts to start the Flow_Table NF would result in errors and prevent it from being started.

Now, to prevent this issue, the Flow_Table NF's initialization sequence has been modified such that it checks to see if the needed ring buffers are already allocated (by a older instance of the Flow_Table NF) and if they are, the NF will free the ring buffers.  Once the memory is sanitized, it proceeds as normal and allocates the memory that it needs. 

Commit log:

* Remove old ring buffers when startup.