[speed_tester] Adds Flag To Control Number of Packets (#235)
Before, the speed_tester NF preallocated 128 packets that it would send to a destination over and over again.  There was no way to throttle the number of allocated packets.

With this change, we add a new optional flag, `-c`, to the speed_tester NF.  Passing a value to this flag controls how many packets the NF will create.  If the flag is not passed, then the NF falls back to its default, 128.  It also defines a max packet limit, currently it is set to `NF_QUEUE_RINGSIZE`.

API Change:

* The macro, `NF_QUEUE_RINGSIZE`, has been moved to `onvm_common.h`

Commit log:

*  [Enhancement #232] Add a -c parameter to specify how many packets will be created when initializing.

    Previously, 128 packets will be created by speed tester. Users have to recompile the code, if
they want to use more or less packets to do the spead test.
    This patch add a -c <packet_number> parameter to the command line of the speed_tester NF. Users can
specify a packet number from 1 to 1M. The defualt packets number is still 128.

* Update the patch according to comments

* Supporting the interaction between -c and -o

* Remove temporary modifications

* move variable definition out of for statement

* Replace tab with spaces

* Move definition of NF_QUEUE_RINGSIZE to onvm_common.h and modify code style according to comments

* Move definition of NF_QUEUE_RINGSIZE to onvm_common.h and modify code style according to comments

* Restore dpdk submodule deleted by mistake