[Bug Fix] LB NF Forwards Source Port MAC Addr (#237)
This change addresses a bug in the Load Balancer NF where two machines connected to a switch could send out packets with the same MAC.  If the NF is sending packets from a client to a backend server, the client sends packets out through the switch with its MAC (11:22:33:44:55:66).  The LB NF receives the packet and then sends it to a backend server while retaining the packet header.  The packet that it would send out would have the same Ethernet frame.  This would cause the switch to panic as two ports are sending packets with the same MAC address.

To resolve this, the Load Balancer is updated to change the MAC address of the packet before it is sent out to a backend server.  The NF grabs the MAC address of its source port that the client sent the packet to, and sets the source MAC address field of the packet to it.  Then, it sends the packet out to the intended backend server. 

Commit log:

* FIX the mac addr for sent out packets