Adds Opcode Processing to ARP NF (#240)
Before, the ARP NF would try to reply to all ARP traffic - even if it was not of type `ARP_OP_REQUEST`.  This is incorrect behavior as it should not attempt to reply to responses and other ARP opcodes.

This change updates the ARP NF packet handler function such that it checks the opcode before attempting to send a response.  Now, the ARP NF will only send a response if the opcode is `ARP_OP_REQUEST`.  If the opcode is `ARP_OP_RESPONSE`, then the NF will simply forward it to the intended recipient or down the NF chain.  If the opcode is anything else defined in DPDK's libraries, then the packet will simply be forwarded down the NF chain.

Commit log:

* Check arp opcode so we don't respond to everything

* Ignore (fwd TONF) unsupported opcodes