Add Latency Measurement to Speed_Tester NF (#256)
This change adds a new, optional, flag to the Speed_Tester NF.  Now, running Speed_Tester with the `-l` flag will generate packets with an additional bit enabled and with a timestamp as data.  The NF will now check for the latency bit when run with the `-l` flag, and then it will extract the timestamp and update the average latency.  The average latency value will be printed with the NF output.

Commit log:

* Initial commit for latency

 - Calculate the latency by timestamping the packets
 - Average the latency over all packets
 - Prints the avg nanosec latency
 * This method is expensive and cuts the throguhput speed *

* working latency in packet

* Clean up latency code

 - Add a flag to enable latency measuring
 - Remove extra testing code

* Readme updates

* Quick README fix