Add ONVM Dockerfile and Update Docker Resources (#38)
This change updates the docker.sh script which helps users launch NFs
inside Docker containers.  It removes an unncessary argument, `-w
WORKINGDIRECTORY`, and changes it to launch containers based off of the
new sdnfv/opennetvm Docker image.

This change also provides a Dockerfile which can be used for docker
image creation.  It creates a docker image based off of Ubuntu 14.04
which is what is uploaded to our [Docker Hub
account](https://hub.docker.com/r/sdnfv/opennetvm/).

With the script changes and use of our Dockerhub account, this change
also updates the Docker README in the docs directory to provide the most
up-to-date information about Docker and OpenNetVM.

Testing:

        * Launched a few containers in various topologies in Docker
containers and ensured traffic flowed through the chain and
latency/throughput matched what was expected when running directly on
the host.