speed tester accepts 0 for initial packets.
However, still defaults to 128 initial packets if the -c option is not specified.
An 'if_c' global variable is added which is changed in the argument
parser function to 1 if the c option is used.

In the main function, it is checked if the if_c variable is set to 1. If
that is not the case, then packet_number is set to the default of 128.