Speed Tester now accepts 0 initial packets (#56)
Enable Speed Tester NF to accept 0 as an argument to the `-c` flag, which determines how many packets to create.

This is useful if you want to use Speed Tester to measure latency for packets generated elsewhere (e.g. real packets could be generated from a different NF, and you can send data through the Speed Tester NF which is configured with 0 packets.  This will allow latency information to be printed out).

Commit log:

* Speed Tester now accepts 0 initial packets

* speed tester accepts 0 for initial packets.

* Changed global variable from 'if_c' to 'use_custom_pkt_count'