Add NFLib Function to Return Packets in Bulk (#269)
Adds support for returning packets in bulk instead of one by one by using  `onvm_nflib_return_pkt_bulk`. Useful for functions that buffer a group of packets before returning them for processing or for NFs that create batches of packets in the fast path.

Currently, NFLib exposes a function that allows an NF to return a packet to the library. When using this function during NF initialization, it doesn't matter how fast you return the packets because it only happens once in the beginning. However, returning packets in the critical path should be done faster. 

Adding a function onvm_nflib_return_pkts should improve the speed for NFs that want to return packets often, such as a load generator. 

__No breaking API changes.__

Commit log:

* Let users return packets in bulk.

* rename bulk function, update comment, style nits

* refactor bulk here as well

* performance enhancement

* Update Releases.md for bulk packet return.

* apply change

Co-Authored-By: rskennedy <rkennedysurf@gmail.com>