Speed Tester Bails if MAC Addr Arg is Invalid (#286)
Argument error checking in the example NFs typically prints a usage message and exits if the user gives an invalid input. However, the error check for MAC address format did not exit after the usage message was printed, making the usage message easy to miss as it is quickly replaced by the stats display of the running speed tester. This bug fix adds a line to return -1 after the usage message is displayed.

Commit Log:

* Exit speed tester on invalid MAC address