Revamps Web Stats and Adds New Events System (#284)
This PR contains changes to the ONVM stats thread, including a new events system which is used for port initialization and NF starting, ready, and stopping. In the future, this should be used for more complex events such as service chain based events, and for core mappings once shared CPU
is completed.

This PR also contains a complete rewrite of the web frontend. The existing code which primarily used jquery has been rewritten and expanded upon in React, using Flow for type checking rather than a full TypeScript implementation. This allows us to maintain application state across pages and to restore graphs to the fully updated state when returning to a graph from a different page.

The core of the new frontend is a small library for streaming stats updates and events to subscribed components, which abstracts much of the work away from the pages and attempts to hide some of the difficulty of maintaining an updated frontend while only polling stats and events files every three seconds. This can be found in [pubsub.js](https://github.com/sdnfv/openNetVM-dev/pull/284/files#diff-3f31cf49194842daab2e0eb805521b6b).

Other new additions are more detailed NF specific information, which can be easily expanded upon by providing the extra info within the C code and using the stats streaming model to simply access the additional information.

A CPU layout page has been templated, but since there are some things still going on with shared CPU / pthreads, I thought it would be best to wait for that to be a bit more clear. It seems the best approach to this might be including core information in the `nf_info` struct.

Please note that **CSV download has been removed** with this update as storing this much ongoing data negatively impacts application performance. This sort of data collection would be best implemented via grepping or some similar functionality from onvm console output.

Commit Log:

* Removing old web stats and updating .gitignore

* Initial react app setup using create-react-app.

* Added flow typing to project (flow.org)

* Removing boilerplate from cra.

* Continuing setup.

* Initial routing setup.

* Using prettier (prettier.io) to format code.

* Added prettier to precommit process.

* Adding tabler-react component library.

* Working on routing and dashboard setup.

* Added working graph example with dynamic updates.

* Routing fixes and data fetching.

* Working on pubsub model for component data handling.

* Refactored stats structure and build basic NF pubsub model.

* Working register/unregister for graphs.

* Added working NF graph.

* Added ability to do list of NFs and fixed resulting bugs.

* Removed debugging statement.

* Updates for single NF view.

* Setting up SingleNFPage with good starting configuration.

* Defining stuff for events setup.

* Added closing of file on cleanup and ignoring generated json.

* Added array creation, file printing, flushing, and initial array creation.

* Fixed comment typo.

* Added events for start, stop, and ready on NFs.

* Added events and events subscription.

* Added ability to catch up on missed events.

* Fixed warnings.

* Started displaying real NF info on right hand panel.

* Fixed react memory leak issue.

* Fixed bug with NFDashboard page and improved dashboard load time.

* Working event list.

* Working on port information.

* Adding extra info and bugfixes for NF dashboard.

* Port graphs now display based on ports being used, no data though.

* Fixed port stats in onvm_stats and fixed frontend data handlers.

* Added coremappingspage framework and added to routes.

- Also updated icons.

* Changed page title.

* Adding legitimate timestamp to events.

* Greatly simplified CORS server and cited source.

* Swapped SimpleHTTPServer in favor of CORS server in case of cross origin.

* Removed RTE_LOG with incorrect info (web stats already started)

* Removed commented out code block.

* Began updating documentation.

* fixing up Usage section.

* Update description of cors server

* Slight wording changes in README

* Removed unused links.

* Wrote web documentation.

* Adding guard for undefined event.

* Fixed method formatting and added check for correct output mode.

* Changed if statement for readability.

* Fixed tabs -> spaces.

* Fixing inverted if statement.

* Changed order of code for NF starting so it has ID

* Checking condition before adding stopped event.

* Adding checks for successful calls for events

* Style fixes.

* Updated the port number for requests.

* Adding frontend serve to start_web_console script.

* Added pre-built web.

* Added script to build web code.

* Added note about versions.

* Fixed hostname issue.

* Updating web build code.

* Exporting another web pid.

* Caused ONVM to fail to start if web fails to start.

* Added checks for whether ports are available.

* Successfully sharing PID variables and exiting web stats.

* Updated style for onvm_nf.c

* Moved cd location in onvm go script.

* Shortened event_msg_buf.

* Style fix to onvm_stats.c

* Added note about refreshing when manager is restarted.