[Bug Fix] ONVM Mgr to Handle Two NF Stop Msgs (#292)
This change fixes a bug to allow the ONVM manager to handle two messages notifying it to stop an NF.  Before, when an NF's status is checked twice in a row, and the state is NF_STOPPED, the ONVM managers NF per service Id count would be corrupted.

Essentially, the order of operations when changing NF info struct's status from NF_RUNNING to NF_STOPPED was incorrect.  Before, when an NF was started, the NF status was set to running and a message with the NF info struct would be put in to the ONVM manager's message queue.  If the NF receives SIGINT immediately after starting, the NFLib cleanup function would set NF state to stopped, and again send a message with the NF info struct to the manager.

As the NF info struct in each message is the same, and if they are next to each other in the queue, updating the state to NF_STOPPED when the signal is received effectively sets the state in both messages to NF_STOPPED.  The NF has not yet been started as the first message has not yet been processed.  The manager will read two messages and try to stop the NF twice.  Doing this causes the NF per service count to be messed up.

This change fixes the bug by changing the order of operations when setting the NF state to NF_STOPPED and how it is checked/handled.  Before, the NF set the state to NF_STOPPED in `onvm_nflib_cleanup()` before it sent a message to the manager, and then manager would try to shut it down if the state was NF_STOPPED.  Now, the NF no longer sets its state in NFLib.  Instead, when an NF receives a signal, it will just send a shutdown message to the manager.  If the NF's state is not NF_SHUTDOWN, then it will change its state and shut it down.

This prevents subsequent shutdown calls for the same NF to enter the shutdown sequence twice.

Closes #288 

Commit log:

* Fix double shutdown bug

* Fix another shutdown bug