Changes Default NF Core From Shared to Dedicated (#309)
NFs are now single core only unless a -s flag is specified, instead of specifying the old -d to reserve a core.

For changing the default behavior to share a code run the NF with -s instead. -s Allows to share cores with other NFs.

If we start 2 NFs in shared core mode they will still occupy 2 cores if 2 cores are available. If this is not the desired behavior use the manual mode (-d flag) to manually pin to desired core

Commit Log:

* Change default to dedicated core mode

* Add readme update

* Fix wrong var name