Make Number of mbufs a Constant Value (#79)
We have been observing performance regression from the last few releases, this PR fixes one of the issues. This PR boosts basic_monitor & pktgen benchmark performance from 7Mpps to 13.1 Mpps. This is done by changing the number of mbufs to a static constant instead of it being a function depending on MAX_NFS number as setting the number of mbufs too high leads to performance degradation.

The best performance is around 16-32 MAX_NFS number. This computes to a num of mbufs value between 27648 (16*1536 + 1536*2) and 52224 (32*1536 + 1536*2). Selected 32767 (2^15 - 1, negative one from DPDK docs)  because it is between those numbers and tends to have the same performance.

Commit log:

* Fix mbuf size to a static constant

* Adjust mbuf num to match dpdk recs